Added utility class for manifest filtering.
