+++ lib/media/preload_manager.js
@@ -9,6 +9,7 @@ goog.provide('shaka.media.PreloadManager');
 goog.require('goog.asserts');
 goog.require('shaka.log');
 goog.require('shaka.util.Error');
+goog.require('shaka.media.ManifestFilterer');
 goog.require('shaka.media.ManifestParser');
 goog.require('shaka.util.PublicPromise');
 goog.require('shaka.media.PreferenceBasedCriteria');
@@ -61,8 +62,8 @@ shaka.media.PreloadManager = class {
     /** @private {function():!shaka.media.DrmEngine} */
     this.createDrmEngine_ = typedPlayerInterface.createDrmEngine;
 
-    /** @private {function(!shaka.extern.Manifest):!Promise} */
-    this.filterManifest_ = typedPlayerInterface.filterManifest;
+    /** @private {!shaka.media.ManifestFilterer} */
+    this.manifestFilterer_ = typedPlayerInterface.manifestFilterer;
 
     /** @private {!shaka.extern.ManifestParser.PlayerInterface} */
     this.manifestPlayerInterface_ =
@@ -170,6 +171,11 @@ shaka.media.PreloadManager = class {
   /** @return {!shaka.util.Stats} */
   getStats() {
     return this.stats_;
+  }
+
+  /** @return {!shaka.media.ManifestFilterer} */
+  getManifestFilterer() {
+    return this.manifestFilterer_;
   }
 
   /**
@@ -302,6 +308,7 @@ shaka.media.PreloadManager = class {
     // const startTime = Date.now() / 1000;
 
     this.drmEngine_ = this.createDrmEngine_();
+    this.manifestFilterer_.setDrmEngine(this.drmEngine_);
 
     this.drmEngine_.configure(this.config_.drm);
 
@@ -311,7 +318,11 @@ shaka.media.PreloadManager = class {
 
     // Now that we have drm information, filter the manifest (again) so that
     // we can ensure we only use variants with the selected key system.
-    await this.filterManifest_(this.manifest_);
+    // const tracksChanged =
+    await this.manifestFilterer_.filterManifest(this.manifest_);
+    // if (tracksChanged) {
+    // TODO: Fire a "shaka.util.FakeEvent.EventName.TracksChanged" event.
+    // }
   }
 
   /**
@@ -329,8 +340,6 @@ shaka.media.PreloadManager = class {
 
     // TODO: Add a method to re-configure the preload manager once it has hooked
     // up to a load object?
-    // TODO: Pass this.currentAdaptationSetCriteria_ into the filterManifest_
-    // method?
 
     // Copy preferred languages from the config again, in case the config was
     // changed between construction and playback.
@@ -456,8 +465,8 @@ shaka.media.PreloadManager = class {
  *   regionTimeline: !shaka.media.RegionTimeline,
  *   qualityObserver: ?shaka.media.QualityObserver,
  *   createDrmEngine: function():!shaka.media.DrmEngine,
- *   filterManifest: function(!shaka.extern.Manifest):!Promise,
- *   networkingEngine: !shaka.net.NetworkingEngine
+ *   networkingEngine: !shaka.net.NetworkingEngine,
+ *   manifestFilterer: !shaka.media.ManifestFilterer
  * }}
  *
  * @property {!shaka.extern.PlayerConfiguration} config
@@ -466,7 +475,7 @@ shaka.media.PreloadManager = class {
  * @property {!shaka.media.RegionTimeline} regionTimeline
  * @property {?shaka.media.QualityObserver} qualityObserver
  * @property {function():!shaka.media.DrmEngine} createDrmEngine
- * @property {function(!shaka.extern.Manifest):!Promise} filterManifest
  * @property {!shaka.net.NetworkingEngine} networkingEngine
+ * @property {!shaka.media.ManifestFilterer} manifestFilterer
  */
 shaka.media.PreloadManager.PlayerInterface;
