+++ lib/player.js
@@ -14,6 +14,7 @@ goog.require('shaka.media.AdaptationSetCriteria');
 goog.require('shaka.media.BufferingObserver');
 goog.require('shaka.media.DrmEngine');
 goog.require('shaka.media.ExampleBasedCriteria');
+goog.require('shaka.media.ManifestFilterer');
 goog.require('shaka.media.ManifestParser');
 goog.require('shaka.media.MediaSourceEngine');
 goog.require('shaka.media.MediaSourcePlayhead');
@@ -670,6 +671,10 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
 
     /** @private {{width: number, height: number}} */
     this.maxHwRes_ = {width: Infinity, height: Infinity};
+
+    /** @private {!shaka.media.ManifestFilterer} */
+    this.manifestFilterer_ = new shaka.media.ManifestFilterer(
+        this.config_, this.maxHwRes_, null);
 
     /** @private {shaka.util.Stats} */
     this.stats_ = null;
@@ -1439,6 +1444,7 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
         // Get manifest and associated values from preloader.
         await mutexWrapOperation(
             () => preloadManager.waitForManifest(), 'waitForManifest');
+        this.manifestFilterer_ = preloadManager.getManifestFilterer();
         this.parserFactory_ = preloadManager.getParserFactory();
         this.parser_ = preloadManager.receiveParser();
         this.regionTimeline_ = preloadManager.receiveRegionTimeline();
@@ -1566,11 +1572,19 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
       // TODO: Mark a flag such that, once this preload manager has attached to
       // a player, we will set disableVideo and filter for audio-only.
     }
+    const manifestFilterer = new shaka.media.ManifestFilterer(
+        config, this.maxHwRes_, null);
     // TODO: Refactor how the player interfaces for manifest and DRM parser work
     // so that they can hook into either the player or preload manager.
     const manifestPlayerInterface = {
       networkingEngine: this.networkingEngine_,
-      filter: (manifest) => this.filterManifest_(manifest),
+      filter: async (manifest) => {
+        const tracksChanged = await manifestFilterer.filterManifest(manifest);
+        if (tracksChanged) {
+          // TODO: The preload manager eat this event too.
+          this.onTracksChanged_();
+        }
+      },
       makeTextStreamsForClosedCaptions: (manifest) => {
         return this.makeTextStreamsForClosedCaptions_(manifest);
       },
@@ -1600,7 +1614,7 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
             this.drmEngine_ ? this.drmEngine_.getDrmInfo() : null;
         // DrmEngine.newInitData() requires mediaKeys to be available.
         if (currentDrmInfo && this.drmEngine_.getMediaKeys()) {
-          this.processDrmInfos_(currentDrmInfo.keySystem, stream);
+          manifestFilterer.processDrmInfos(currentDrmInfo.keySystem, stream);
         }
       },
       onManifestUpdated: () => {
@@ -1673,12 +1687,9 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
       },
     };
     // TODO: Make this based on the saved config, not the current config.
+    /** @return {!shaka.media.DrmEngine} */
     const createDrmEngine = () => {
       return this.createDrmEngine(drmPlayerInterface);
-    };
-    // TODO: Filter this based on the saved config, not the current config.
-    const filterManifest = async (manifest) => {
-      await this.filterManifest_(manifest);
     };
     /** @type {!shaka.media.PreloadManager.PlayerInterface} */
     const playerInterface = {
@@ -1687,7 +1698,7 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
       regionTimeline,
       qualityObserver,
       createDrmEngine,
-      filterManifest,
+      manifestFilterer,
       networkingEngine: this.networkingEngine_,
     };
     return new shaka.media.PreloadManager(
@@ -2978,6 +2989,8 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
    * @private
    */
   applyConfig_() {
+    this.manifestFilterer_ = new shaka.media.ManifestFilterer(
+        this.config_, this.maxHwRes_, this.drmEngine_);
     if (this.parser_) {
       const manifestConfig =
           shaka.util.ObjectUtils.cloneObject(this.config_.manifest);
@@ -2994,9 +3007,14 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
       this.streamingEngine_.configure(this.config_.streaming);
 
       // Need to apply the restrictions.
+      // this.filterManifestWithRestrictions_() may throw.
       try {
-        // this.filterManifestWithRestrictions_() may throw.
-        this.filterManifestWithRestrictions_(this.manifest_);
+        if (this.loadMode_ != shaka.Player.LoadMode.DESTROYED) {
+          if (this.manifestFilterer_.filterManifestWithRestrictions(
+              this.manifest_)) {
+            this.onTracksChanged_();
+          }
+        }
       } catch (error) {
         this.onError_(error);
       }
@@ -5260,82 +5278,6 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
   }
 
   /**
-   * Filters a manifest, removing unplayable streams/variants.
-   *
-   * @param {?shaka.extern.Manifest} manifest
-   * @private
-   */
-  async filterManifest_(manifest) {
-    await this.filterManifestWithStreamUtils_(manifest);
-    this.filterManifestWithRestrictions_(manifest);
-  }
-
-  /**
-   * Filters a manifest, removing unplayable streams/variants.
-   *
-   * @param {?shaka.extern.Manifest} manifest
-   * @private
-   */
-  async filterManifestWithStreamUtils_(manifest) {
-    goog.asserts.assert(manifest, 'Manifest should exist!');
-    await shaka.util.StreamUtils.filterManifest(this.drmEngine_, manifest);
-    this.checkPlayableVariants_(manifest);
-  }
-
-
-  /**
-   * Apply the restrictions configuration to the manifest, and check if there's
-   * a variant that meets the restrictions.
-   *
-   * @param {?shaka.extern.Manifest} manifest
-   * @private
-   */
-  filterManifestWithRestrictions_(manifest) {
-    // Return if |destroy| is called.
-    if (this.loadMode_ == shaka.Player.LoadMode.DESTROYED) {
-      return;
-    }
-    const tracksChanged = shaka.util.StreamUtils.applyRestrictions(
-        manifest.variants, this.config_.restrictions, this.maxHwRes_);
-    if (tracksChanged && this.streamingEngine_) {
-      this.onTracksChanged_();
-    }
-
-    // We may need to create new sessions for any new init data.
-    const currentDrmInfo =
-        this.drmEngine_ ? this.drmEngine_.getDrmInfo() : null;
-    // DrmEngine.newInitData() requires mediaKeys to be available.
-    if (currentDrmInfo && this.drmEngine_.getMediaKeys()) {
-      for (const variant of manifest.variants) {
-        this.processDrmInfos_(currentDrmInfo.keySystem, variant.video);
-        this.processDrmInfos_(currentDrmInfo.keySystem, variant.audio);
-      }
-    }
-    this.checkRestrictedVariants_(manifest);
-  }
-
-  /**
-   * @param {string} keySystem
-   * @param {?shaka.extern.Stream} stream
-   * @private
-   */
-  processDrmInfos_(keySystem, stream) {
-    if (!stream) {
-      return;
-    }
-
-    for (const drmInfo of stream.drmInfos) {
-      // Ignore any data for different key systems.
-      if (drmInfo.keySystem == keySystem) {
-        for (const initData of (drmInfo.initData || [])) {
-          this.drmEngine_.newInitData(
-              initData.initDataType, initData.initData);
-        }
-      }
-    }
-  }
-
-  /**
    * @param {shaka.extern.Variant} initialVariant
    * @param {number} time
    * @return {!Promise.<number>}
@@ -5651,7 +5593,7 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
   updateAbrManagerVariants_() {
     try {
       goog.asserts.assert(this.manifest_, 'Manifest should exist by now!');
-      this.checkRestrictedVariants_(this.manifest_);
+      this.manifestFilterer_.checkRestrictedVariants(this.manifest_);
     } catch (e) {
       this.onError_(e);
       return false;
@@ -6351,7 +6293,7 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
           'restrictions so we don\'t select those tracks.');
     }
 
-    const restrictedStatuses = shaka.Player.restrictedStatuses_;
+    const restrictedStatuses = shaka.media.ManifestFilterer.restrictedStatuses;
     let tracksChanged = false;
 
     // Only filter tracks for keys if we have some key statuses to look at.
@@ -6455,93 +6397,6 @@ shaka.Player = class extends shaka.util.FakeEventTarget {
       } else {
         timeline.setDuration(playRangeEnd);
       }
-    }
-  }
-
-  /**
-   * Checks if the variants are all restricted, and throw an appropriate
-   * exception if so.
-   *
-   * @param {shaka.extern.Manifest} manifest
-   *
-   * @private
-   */
-  checkRestrictedVariants_(manifest) {
-    const restrictedStatuses = shaka.Player.restrictedStatuses_;
-    const keyStatusMap =
-        this.drmEngine_ ? this.drmEngine_.getKeyStatuses() : {};
-    const keyIds = Object.keys(keyStatusMap);
-    const isGlobalStatus = keyIds.length && keyIds[0] == '00';
-
-    let hasPlayable = false;
-    let hasAppRestrictions = false;
-
-    /** @type {!Set.<string>} */
-    const missingKeys = new Set();
-
-    /** @type {!Set.<string>} */
-    const badKeyStatuses = new Set();
-
-    for (const variant of manifest.variants) {
-      // TODO: Combine with onKeyStatus_.
-      const streams = [];
-      if (variant.audio) {
-        streams.push(variant.audio);
-      }
-      if (variant.video) {
-        streams.push(variant.video);
-      }
-
-      for (const stream of streams) {
-        if (stream.keyIds.size) {
-          for (const keyId of stream.keyIds) {
-            const keyStatus = keyStatusMap[isGlobalStatus ? '00' : keyId];
-            if (!keyStatus) {
-              missingKeys.add(keyId);
-            } else if (restrictedStatuses.includes(keyStatus)) {
-              badKeyStatuses.add(keyStatus);
-            }
-          }
-        }  // if (stream.keyIds.size)
-      }
-
-      if (!variant.allowedByApplication) {
-        hasAppRestrictions = true;
-      } else if (variant.allowedByKeySystem) {
-        hasPlayable = true;
-      }
-    }
-
-    if (!hasPlayable) {
-      /** @type {shaka.extern.RestrictionInfo} */
-      const data = {
-        hasAppRestrictions,
-        missingKeys: Array.from(missingKeys),
-        restrictedKeyStatuses: Array.from(badKeyStatuses),
-      };
-      throw new shaka.util.Error(
-          shaka.util.Error.Severity.CRITICAL,
-          shaka.util.Error.Category.MANIFEST,
-          shaka.util.Error.Code.RESTRICTIONS_CANNOT_BE_MET,
-          data);
-    }
-  }
-
-  /**
-   * Confirm some variants are playable. Otherwise, throw an exception.
-   * @param {!shaka.extern.Manifest} manifest
-   * @private
-   */
-  checkPlayableVariants_(manifest) {
-    const valid = manifest.variants.some(shaka.util.StreamUtils.isPlayable);
-
-    // If none of the variants are playable, throw
-    // CONTENT_UNSUPPORTED_BY_BROWSER.
-    if (!valid) {
-      throw new shaka.util.Error(
-          shaka.util.Error.Severity.CRITICAL,
-          shaka.util.Error.Category.MANIFEST,
-          shaka.util.Error.Code.CONTENT_UNSUPPORTED_BY_BROWSER);
     }
   }
 
@@ -6782,18 +6637,6 @@ shaka.Player.version = 'v4.5.0' + '-uncompiled';  // x-release-please-version
 // Initialize the deprecation system using the version string we just set
 // on the player.
 shaka.Deprecate.init(shaka.Player.version);
-
-
-/**
- * These are the EME key statuses that represent restricted playback.
- * 'usable', 'released', 'output-downscaled', 'status-pending' are statuses
- * of the usable keys.  'expired' status is being handled separately in
- * DrmEngine.
- *
- * @const {!Array.<string>}
- * @private
- */
-shaka.Player.restrictedStatuses_ = ['output-restricted', 'internal-error'];
 
 
 /** @private {!Object.<string, function():*>} */
